Refactored package names and file names, updated mapping to new file names, fixed broken path moving js files to static folder
