+++ mgl-webapp/src/main/java/com/organization/mvcproject/model/Company.java
@@ -1,19 +1,16 @@
 
-package com.organization.mvcproject.MGL_Task1.model;
+package com.organization.mvcproject.model;
 
-import java.util.ArrayList;
 import java.util.List;
 
-import org.springframework.stereotype.Component;
 
-//TODO 1.0 remove @Component annotation, it is not used
-@Component
+
 public class Company {
 
 	private Long id;
 	private String name;
-	//TODO 1.0 this List doesn't need to be initialized at declaration. 
-	private List<Game> gamesMade = new ArrayList<Game>();
+	 
+	private List<Game> gamesMade;
 
 	public Long getId() {
 		return id;
