+++ mgl-webapp/src/main/java/com/organization/mvcproject/service/GameService.java
@@ -1,0 +1,13 @@
+package com.organization.mvcproject.service;
+
+import java.util.List;
+
+import com.organization.mvcproject.model.Game;
+
+public interface GameService {
+
+	List<Game> retrieveAllGames();
+
+	Game saveGame(Game game);
+
+}
