+++ mgl-webapp/src/main/webapp/WEB-INF/views/reviewCreatePage.jsp
@@ -7,8 +7,8 @@
         <head>
         <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.7.8/angular.min.js"></script>
         <script src="resources/static/js/app.js" /></script>
-        <script src="resources/static/js/service/MGL_Task1.service.js"></script>
-        <script src="resources/static/js/controller/MGL_Task1.controller.js"></script>
+        <script src="resources/static/js/service.js"></script>
+        <script src="resources/static/js/controller.js"></script>
             <!-- Bootstrap CSS -->
             <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
 
@@ -27,15 +27,15 @@
            <mgl:myNav/>
             <div id="ReviewFormDiv" class="container">
                 <br>
-                <form name="submitReviewForm" method="POST" action="addReview">
+                <form name="submitReviewForm" method="POST" action="/addReview">
                     <table class="table table-dark text-light">
                         <tr>
                             <td><label>Review Body*</label></td>
-                            <td><textarea name="reviewBody" class="form-control" placeholder="What did you like/dislike about the game? [required]" required></textarea></td>
+                            <td><textarea name="body" class="form-control" placeholder="What did you like/dislike about the game? [required]" required></textarea></td>
                         </tr>
                         <tr>
                             <td><label>Author</label></td>
-                            <td><input type="text" name="author" class="form-control" placeholder="Your name (leave black to leave an anonymous review)"></input>
+                            <td><input type="text" name="author" class="form-control" placeholder="Your name (leave blank to leave an anonymous review)"></input>
                             </td>
                         </tr>
                         <tr>
