--- mgl-webapp/src/main/webapp/resources/static/js/MGL_Task1.controller.js
@@ -1,26 +1,0 @@
-'use strict';
-
-angular.module('MGL_Task1_app').controller('MGL_Task1_Controller',
-		[ 'MGL_Task1_Service', function(MGL_Task1_Service) {
-			var self = this;
-			self.game = {
-				game_id : '',
-				game_name : '',
-				game_genre : ''
-			};
-			self.games = [];
-
-			self.fetchAllGames = function(){
-				MGL_Task1_Service.fetchAllGames().then(function(data) {
-					self.games = data;
-				});
-			}
-
-			self.addGame = function(){
-				return MGL_Task1_Service.createGame(self.game).then( function() {
-				self.fetchAllGames();
-				});
-			}
-
-			self.fetchAllGames();
-		} ]);
