+++ mgl-webapp/src/main/webapp/resources/static/js/app.js
@@ -1,3 +1,3 @@
 'use strict';
 
-angular.module('MGL_Task1_app',[]);
+angular.module('GameApp',[]);
