+++ mgl-webapp/src/main/webapp/resources/static/js/controller.js
@@ -1,0 +1,26 @@
+'use strict';
+
+angular.module('GameApp').controller('GameController',
+		[ 'GameService', function(GameService) {
+			var self = this;
+			self.game = {
+				id : '',
+				name : '',
+				genre : ''
+			};
+			self.games = [];
+
+			self.fetchAllGames = function(){
+				GameService.fetchAllGames().then(function(data) {
+					self.games = data;
+				});
+			}
+
+			self.addGame = function(){
+				return GameService.createGame(self.game).then( function() {
+				self.fetchAllGames();
+				});
+			}
+
+			self.fetchAllGames();
+		} ]);
