+++ mgl-webapp/src/main/webapp/resources/static/js/service.js
@@ -1,6 +1,6 @@
 'use strict';
 
-angular.module('MGL_Task1_app').factory('MGL_Task1_Service', ['$http', function($http) {
+angular.module('GameApp').factory('GameService', ['$http', function($http) {
 
 		var REST_SERVICE_URI = 'game/';
 
@@ -12,14 +12,14 @@ angular.module('MGL_Task1_app').factory('MGL_Task1_Service', ['$http', function(
 		return factory;
 
 		function fetchAllGames() {
-			return $http.get(REST_SERVICE_URI + 'getAll').then(function(response) {
+			return $http.get(REST_SERVICE_URI).then(function(response) {
 					return response.data;
 				}
 			);
 		}
 
 		function createGame(game) {
-			return $http.post(REST_SERVICE_URI + 'createGame', game).then(function(response) {
+			return $http.post(REST_SERVICE_URI, game).then(function(response) {
 					return response.data;
 				}
 			);
